Refactoring price formatting
