+++ app/shopping.py
@@ -2,16 +2,7 @@ import os
 from datetime import datetime
 from pandas import read_csv
 
-def format_usd(my_price):
-    """
-    Formats a number as USD with dollar sign and two decimals (and also thousands separator)
-
-    Params my_price is a number (int or float) that we want to format
-
-    Examples: format_usd(10)
-
-    """
-    return f"${my_price:,.2f}"
+from app.number_decorators import format_usd
 
 def find_product(product_id, all_products):
     """
